refactor: extract media handlers and create shared media utilities
Create huf/ai/media_utils.py with shared helpers that eliminate ~200
lines of duplicated boilerplate across image/audio/OCR handlers:
- get_agent_provider_config(): load agent, provider, API key
- get_next_conversation_index(): sequential message ordering
- create_agent_message(): Agent Message document creation
- save_media_file(): file save with message/conversation attachment
- emit_message_event(): realtime socket event publishing
- update_conversation_total_messages(): conversation counter update

Extract all media handlers to huf/ai/handlers/media.py:
- handle_generate_image (image generation via LiteLLM)
- handle_ocr_document (OCR via LiteLLM OCR/vision endpoints)
- handle_generate_audio (TTS via LiteLLM speech)
- handle_transcribe_audio (STT via LiteLLM transcription)
- All associated helper/config functions (_resolve_tts_config, etc.)

Also fixes: replace print() with frappe.logger().debug() in image
generation (was print("Returned images: ", images)).

Backward-compatible re-exports kept in sdk_tools.py. The file is now
reduced from 2,682 lines to 465 lines.
